+++ src/object-store/src/object_store.cpp
@@ -171,7 +171,7 @@ void ObjectStore::verify_schema(Schema const& actual_schema, Schema& target_sche
         errors.insert(errors.end(), more_errors.begin(), more_errors.end());
     }
     if (errors.size()) {
-        throw SchemaUpdateValidationException(errors);
+        throw SchemaMismatchException(errors);
     }
 }
 
@@ -224,25 +224,25 @@ static void copy_property_values(const Property& old_property, const Property& n
 
 static void copy_property_values(const Property& source, const Property& destination, Table& table) {
     switch (destination.type) {
-        case PropertyTypeInt:
+        case PropertyType::Int:
             copy_property_values(source, destination, table, &Table::get_int, &Table::set_int);
             break;
-        case PropertyTypeBool:
+        case PropertyType::Bool:
             copy_property_values(source, destination, table, &Table::get_bool, &Table::set_bool);
             break;
-        case PropertyTypeFloat:
+        case PropertyType::Float:
             copy_property_values(source, destination, table, &Table::get_float, &Table::set_float);
             break;
-        case PropertyTypeDouble:
+        case PropertyType::Double:
             copy_property_values(source, destination, table, &Table::get_double, &Table::set_double);
             break;
-        case PropertyTypeString:
+        case PropertyType::String:
             copy_property_values(source, destination, table, &Table::get_string, &Table::set_string);
             break;
-        case PropertyTypeData:
+        case PropertyType::Data:
             copy_property_values(source, destination, table, &Table::get_binary, &Table::set_binary);
             break;
-        case PropertyTypeDate:
+        case PropertyType::Date:
             copy_property_values(source, destination, table, &Table::get_timestamp, &Table::set_timestamp);
             break;
         default:
@@ -319,8 +319,8 @@ void ObjectStore::create_tables(Group *group, Schema &target_schema, bool update
             if (!current_prop || current_prop->table_column == npos) {
                 switch (target_prop.type) {
                         // for objects and arrays, we have to specify target table
-                    case PropertyTypeObject:
-                    case PropertyTypeArray: {
+                    case PropertyType::Object:
+                    case PropertyType::Array: {
                         TableRef link_table = ObjectStore::table_for_object_type(group, target_prop.object_type);
                         REALM_ASSERT(link_table);
                         target_prop.table_column = table->add_column_link(DataType(target_prop.type), target_prop.name, *link_table);
@@ -524,17 +524,16 @@ DuplicatePrimaryKeyValueException::DuplicatePrimaryKeyValueException(std::string
 SchemaValidationException::SchemaValidationException(std::vector<ObjectSchemaValidationException> const& errors) :
     m_validation_errors(errors)
 {
-    m_what ="The following errors were encountered during schema validation: ";
+    m_what = "Schema validation failed due to the following errors: ";
     for (auto const& error : errors) {
         m_what += std::string("\n- ") + error.what();
     }
 }
 
-
-SchemaUpdateValidationException::SchemaUpdateValidationException(std::vector<ObjectSchemaValidationException> const& errors) :
-    SchemaValidationException(errors)
+SchemaMismatchException::SchemaMismatchException(std::vector<ObjectSchemaValidationException> const& errors) :
+m_validation_errors(errors)
 {
-    m_what ="Migration is required due to the following errors: ";
+    m_what = "Migration is required due to the following errors: ";
     for (auto const& error : errors) {
         m_what += std::string("\n- ") + error.what();
     }
@@ -561,7 +560,7 @@ MissingPropertyException::MissingPropertyException(std::string const& object_typ
 InvalidNullabilityException::InvalidNullabilityException(std::string const& object_type, Property const& property) :
     ObjectSchemaPropertyException(object_type, property)
 {
-    if (property.type == PropertyTypeObject) {
+    if (property.type == PropertyType::Object) {
         m_what = "'Object' property '" + property.name + "' must be nullable.";
     }
     else {